Finished All Foreign Key Migration Relationships

+++ db/migrate/20190611094054_create_encounters.rb
@@ -3,7 +3,7 @@ class CreateEncounters < ActiveRecord::Migration[5.2]
     create_table   :encounters, :primary_key => :encounter_id do |t|
       t.integer        :encounter_type_id
     	t.integer       :program_id
-    	t.bigint      :person_id
+    	t.integer      :person_id
     	t.datetime   :visit_date
     	t.boolean    :voided, null: false, default: 0
       t.bigint        :voided_by
